Separated DB code
Moved code that handles database to another file in package main.
